Add files via upload
The code is written in Python and uses various libraries such as scikit-learn and pandas for data processing and model training. The data used to train and test the model is included in the repository, along with Jupyter notebooks that demonstrate the model's performance.

This project has real-world applications in healthcare, as it could potentially be used by healthcare professionals to detect the disease early and provide better patient care. It could also be integrated into health tracking apps and wearables to provide personalized diabetes risk assessments for individuals.

Feel free to clone this repository and use the code and data for your own research or projects. If you have any questions or suggestions, please don't hesitate to open an issue or reach out to me directly.